past_projects/Allphin_Alex/Compiled_Code/requirements.txt
tensorflow == 2.2.0
numpy
matplotlib
nibabel
tqdm
pydicom
sklearn
keras-unet
tifffile
pillow
scikit-image
tensorflow-addons == 0.11.0
jupyter